# specs/001-resume-upload-rating/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Resume Fit Evaluator API
  version: 0.3.0
  description: |
    Single-user FastAPI service. Maintains a library of baseline resume
    profiles, evaluates job descriptions against a selected profile with
    three-tier scoring, and tracks history. Uploaded PDFs are persisted on
    local disk under `${STORAGE_DIR}/profiles/<id>/<utc-ts>.pdf` and the
    absolute path is returned on the response.
    All responses follow the envelope `{ data, error, meta }`.
servers:
  - url: http://localhost:8000
    description: Local dev
...
paths:
...
  /api/jobs/{job_id}:
    get:
      tags: [jobs]
      summary: Get durable tailoring job status
      description: |
        Returns the current state for a durable tailoring job created by
        evaluation. Clients poll this endpoint until `status` is `succeeded`,
        `failed`, or `cancelled`. On `succeeded`, fetch
        `/api/history/{job_analysis_id}` for the generated resume.
      parameters:
        - in: path
          name: job_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Durable job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TailoringJobEnvelope"
        "404":
          $ref: "#/components/responses/Error"
    post:
      tags: [jobs]
      summary: Request cancellation for a durable AI job
      description: |
        Cancels queued or retry-wait jobs immediately. Running jobs move to
        `cancel_requested`; workers observe that signal before committing
        generated output.
      parameters:
        - in: path
          name: job_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Updated durable job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TailoringJobEnvelope"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  schemas:
    Meta:
      type: object
      properties:
        request_id: { type: string }
        cached: { type: boolean }
      additionalProperties: true
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    TailoringJob:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind:
          type: string
          enum: [tailor, evaluate, evaluate_bulk, evaluate_listing, evaluate_pending_listings]
        job_analysis_id: { type: string, format: uuid, nullable: true }
        parent_job_id: { type: string, format: uuid, nullable: true }
        dedupe_key: { type: string, nullable: true }
        priority: { type: integer }
        run_after: { type: string, format: date-time }
        attempts: { type: integer }
        max_attempts: { type: integer }
        locked_by: { type: string, nullable: true }
        locked_at: { type: string, format: date-time, nullable: true }
        status:
          type: string
          enum: [queued, running, retry_wait, cancel_requested, cancelled, succeeded, failed]
        profile_id: { type: integer, nullable: true }
        job_listing_id: { type: string, format: uuid, nullable: true }
        generated_resume_id: { type: string, format: uuid, nullable: true }
        prompt_version: { type: string, nullable: true }
        request_id: { type: string, nullable: true }
        input_payload:
          type: object
          additionalProperties: true
        result_payload:
          type: object
          nullable: true
          additionalProperties: true
        error_code: { type: string, nullable: true }
        error_message: { type: string, nullable: true }
        progress_current: { type: integer }
        progress_total: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
...
    TailoringJobEnvelope:
      type: object
      properties:
        data: { $ref: "#/components/schemas/TailoringJob" }
        error: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
...
    ErrorEnvelope:
      type: object
      properties:
        data: { type: "null" }
        error: { $ref: "#/components/schemas/Error" }
        meta: { $ref: "#/components/schemas/Meta" }
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"